android: extract HottestPosts composable
